fix(apidata.py-VoxcoDataGrabber): fix bug where integer showed for fill
Fixed a bug where an integer such as '1' would show where a fill was supposed to be. If statements have been put in place to check if the text can be changed to an integer, and if so, the fill will be pulled from a fill block instead.

BREAKING CHANGE: This change my break how the web projects fill. It may be possible that computed variables get passed over when questions are created, but this has not yet been confirmed to happen. The issue where an integer shows for mix mode projects has been resolved.