fix(server->dataManagement.py): fix zip calculation
Fixed a bug where zip calculation wasnt wrking properly. This was due to the files using vtype and stype in their criteria. Criterion need to match in order to properly merge. Replace statements were used to filter the text and make the proper criterion match.

BREAKING CHANGE: This may cause other criterion to merge data in a way that is undesireable in other areas, this is not yet confirmed.